requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
requests==2.31.0
schedule==1.2.1
python-telegram-bot==20.8
openai==1.12.0
python-dotenv==1.0.1
firecrawl-py==1.13.5
gunicorn==21.2.0
beautifulsoup4==4.12.3
pytz==2024.1